requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
SpeechRecognition==3.10.0
gTTS==2.3.2
PyAudio==0.2.13
openai==0.28.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-multipart==0.0.6
pydantic==2.3.0
pytest==7.4.2
httpx==0.25.0